+++ tests/testthat/test-offset.R
@@ -31,23 +31,24 @@ test_that("verbose", {
   expect_equal(out2$Mean, c(256.42016, 289.02697, 707.83022), tolerance = 1e-3)
 
   expect_message(estimate_means(moff, "x"), regex = "We also found")
-  expect_message(
-    estimate_means(moff, "x", estimate = "average"),
-    regex = "(e.g., \"population\")",
-    fixed = TRUE
-  )
 
   expect_silent({
     out1 <- estimate_means(moff, "x", offset = 100)
   })
-  expect_message(
-    {
-      out2 <- estimate_means(moff, "x", estimate = "average", offset = 100)
-    },
-    regex = "For"
-  )
+  expect_silent({
+    out2 <- estimate_means(moff, "x", estimate = "average", offset = 100)
+  })
   expect_equal(out1$Mean, c(664.68547, 1023.27456, 1474.41949), tolerance = 1e-3)
-  expect_equal(out2$Mean, c(256.42016, 289.02697, 707.83022), tolerance = 1e-3)
+  expect_equal(out2$Mean, c(664.68547, 1023.27456, 1474.41949), tolerance = 1e-3)
+
+  # validate that offset works for estimate = "average"
+  out1 <- marginaleffects::avg_predictions(
+    moff,
+    variables = list(offset_1 = 100),
+    by = "x"
+  )
+  out2 <- estimate_means(moff, "x", estimate = "average", offset = 100)
+  expect_equal(out1$estimate, out2$Mean, tolerance = 1e-4)
 
   # validate that offset works for estimate = "population"
   out1 <- marginaleffects::avg_predictions(